more loading work, updating pagination
